feat: rename logo-request-logs file and update schema import path
